test: refactor test utils
